requirements.txt
schedule
requests
pyyaml
deepmerge
python-telegram-bot
bs4
